完成分页功能

+++ zihao-service/entity/dto/PageDto.go
@@ -4,8 +4,8 @@ package dto
 页面对象
  */
 type PageDto struct {
-	Page int `json:"page"`
-	Row int `json:"row"`
-	Records int `json:"records"`
-	Total int `json:"total"`
+	Page int64 `json:"page"`
+	Row int64 `json:"row"`
+	Records int64 `json:"records"`
+	Total int64 `json:"total"`
 }
